Redo the way clicks on the graph preview pane are handled using an event filter.
Instigation: There was a bug causing the Options menu to get reset every time
I click on the graph, making the "Use default theme" menu item get replaced
with "Use dark theme".  This commit fixes that bug.

The bug was ultimately caused by clicks on the plot being sent to the main
window's preview_pane_clicked slot, which would call receive_widget, which
would re-set up the menu.

To fix the immediate bug, I made setup_menu_bar return the existing menu
bar if it was alredy created.

But it's a waste of resource and it's confusing that preview_pane_clicked
was being called while the graph window was open.  So redid preview_pane_clicked
to not call receive widget if it doesn't need to.

Also, to avoid confusion, I renamed preview_pane_clicked to open_graph_window.

Also, use an event filter to intercept clicks to the custom plot, so the custom
plot does not have to check its preview_mode variable when processing clicks.
(That last change is kind of debatable, we could change it back later.)