work in progress: Try making the manual target box always be visible.
We probably don't want the box to be separabable from the "Stop motor"
button which people might need to shut down their system if something
bad is happening.  So it doesn't make that much sense to put it in its
own window (unless we are going to copy the 'Stop motor' button).