gui: Make unimplemented settings invisible instead of disabled.
This should provide a simpler, less confusing interface for umc06a users.

It also lets them make the window a little bit smaller.

And I tested it to make sure that Qt automatically grows the window and doesn't
do anything awful when switching from a umc06a to a different Jrk.